Join local storage functions in one file
