--- src/lib/loadLocally.js
@@ -1,8 +1,0 @@
-export default function loadLocally(key) {
-  const jsonString = localStorage.getItem(key)
-  try {
-    return JSON.parse(jsonString)
-  } catch (error) {
-    console.log(error)
-  }
-}
