--- src/lib/removeLocally.js
@@ -1,3 +1,0 @@
-export default function removeLocally(key) {
-  localStorage.removeItem(key)
-}
