--- src/lib/saveLocally.js
@@ -1,4 +1,0 @@
-export default function saveLocally(key, data) {
-  const jsonString = JSON.stringify(data)
-  localStorage.setItem(key, jsonString)
-}
